+++ test/governer-test.ts
@@ -51,13 +51,6 @@ describe("Governor Tests", function () {
       const instance = await upgrades.deployProxy(Governor, args);
       await instance.deployed();
     });
-
-    it("Verify if the GODHolder contract is upgradeable safe ", async function () {
-      const Governor = await ethers.getContractFactory("GODHolder");
-      const args = [zeroAddress, zeroAddress];
-      const instance = await upgrades.deployProxy(Governor, args);
-      await instance.deployed();
-    });
   });
 
   async function deployFixture() {
@@ -427,37 +420,6 @@ describe("Governor Tests", function () {
           zeroAddress
         )
       ).to.revertedWith("Initializable: contract is already initialized");
-    });
-
-    it("Verify GODHolder initialize should be failed for initialize called after instance created", async function () {
-      const { godHolder } = await loadFixture(deployFixture);
-      await expect(
-        godHolder.initialize(zeroAddress, zeroAddress)
-      ).to.revertedWith("Initializable: contract is already initialized");
-    });
-
-    it("Verify GODHolder grabtoken revert", async function () {
-      const { godHolder, signers, mockBaseHTS } = await loadFixture(
-        deployFixture
-      );
-      await mockBaseHTS.setPassTransactionCount(1);
-      await expect(
-        godHolder.grabTokensFromUser(signers[0].address)
-      ).to.revertedWith("GODHolder: token transfer failed to contract.");
-    });
-
-    it("Verify GODHolder revertTokensForVoter revert", async function () {
-      const { godHolder, signers, tokenCont } = await loadFixture(
-        deployFixture
-      );
-      await expect(godHolder.revertTokensForVoter()).to.revertedWith(
-        "GODHolder: No amount for the Voter."
-      );
-      godHolder.grabTokensFromUser(signers[0].address);
-      await tokenCont.setTransaferFailed(true);
-      await expect(godHolder.revertTokensForVoter()).to.revertedWith(
-        "GODHolder: token transfer failed from contract."
-      );
     });
 
     it("Verify cancel proposal flow when proposal not found", async function () {
